#108 moved main buttons
