+++ Frontend/src/app/dashboard/main-buttons/main-buttons.component.css
@@ -1,0 +1,60 @@
+.Button-wrapper {
+    margin-left: 5vw;
+    margin-right: 5vw;
+    margin-top: 10vh;
+    height: fit-content;
+    height: 60vh;
+}
+
+
+
+.auswahlbutton {
+    display: block;
+    width: 100%;
+    height: 20vh;
+    position: relative;
+    background-color: var(--asta-blue);
+    border-top-left-radius: 20px;
+    border-top-right-radius: 20px;
+    text-align: center;
+}
+
+.subbutton {
+    display: block;
+    width: 50%;
+    height: 3vw;
+    float: inline-start;
+    background-color: var(--asta-bluelight);
+    color: var(--asta-white);
+}
+
+.sbright {
+    border-bottom-right-radius: 20px;
+}
+
+.sbleft {
+    border-bottom-left-radius: 20px;
+}
+
+.subbutton-wrapper {
+    display: block;
+    position: relative;
+    margin-left: 5%;
+    margin-right: 5%;
+    float: inline-start;
+    margin-bottom: 5%;
+    height: fit-content;
+    width: 20%;
+    height: 20vh;
+}
+
+h1 {
+    font-size: 3vmin;
+    text-decoration: underline;
+    color:var(--asta-white);
+}
+
+h2 {
+    font-size: 2vmin;
+    color:var(--asta-white);
+}
