--- Frontend/src/app/main-buttons/main-buttons.component.html
@@ -1,16 +1,0 @@
-<div class="Button-wrapper">
-    @for(t of this.templateService.getAntragsTemplates(); track t) {
-        <div class="subbutton-wrapper">
-            <div class="auswahlbutton">
-                <h1>{{t.titel}}</h1>
-                <h2>{{t.description}}</h2>
-            </div>
-            <button type="button" class="subbutton sbleft" routerLink="/Form/{{t.id}}">Antrag</button>
-            @for(a of this.templateService.getAbrechnungsTemplates(); track a) {
-                @if(a.titel == t.titel) {
-                    <button type="button" class="subbutton sbright" routerLink="/Form/{{a.id}}">Abrechnung</button>
-                }
-            }
-        </div>
-    }
-</div>
